// src/types.d.ts
export type TPositionInfo = {
  line: number
  column: number
}
// ...
export type TSpecialNodeMap = {
  [ENodeType.DictionaryProperty]: IDictionaryProperty
  [ENodeType.AssignmentParam]: IAssignmentParam
  [ENodeType.ExceptHandler]: IExceptHandler
}

/** 表达式节点映射表 */
export type TExpressionNodeMap = {
  [ENodeType.NoneLiteral]: INoneLiteral
  [ENodeType.BooleanLiteral]: IBooleanLiteral
  [ENodeType.NumberLiteral]: INumberLiteral
  [ENodeType.StringLiteral]: IStringLiteral
  [ENodeType.Identifier]: IIdentifier
  [ENodeType.UnaryExpression]: IUnaryExpression
  [ENodeType.IfExpression]: IIfExpression
  [ENodeType.ArrayExpression]: IArrayExpression
  [ENodeType.DictionaryExpression]: IDictionaryExpression
  [ENodeType.BinaryExpression]: IBinaryExpression
  [ENodeType.VariableDeclaration]: IVariableDeclaration
  [ENodeType.AssignmentExpression]: IAssignmentExpression
  [ENodeType.SliceExpression]: ISliceExpression
  [ENodeType.MemberExpression]: IMemberExpression
  [ENodeType.CallExpression]: ICallExpression
  [ENodeType.TupleExpression]: ITupleExpression
}

export type TExpressionNode<T extends keyof TExpressionNodeMap = keyof TExpressionNodeMap> = TExpressionNodeMap[T]

/** 语句节点映射表 */
export type TStatementNodeMap = {
  [ENodeType.ImportStatement]: IImportStatement
  [ENodeType.FunctionDeclaration]: IFunctionDeclaration
  [ENodeType.ClassDeclaration]: IClassDeclaration
  [ENodeType.BlockStatement]: IBlockStatement
  [ENodeType.EmptyStatement]: IEmptyStatement
  [ENodeType.IfStatement]: IIfStatement
  [ENodeType.TryStatement]: ITryStatement
  [ENodeType.Program]: IProgram
}

export type TStatementNode<T extends keyof TStatementNodeMap = keyof TStatementNodeMap> = TStatementNodeMap[T]

/** 所有节点的映射表 */
export type TNodeMap = TSpecialNodeMap & TExpressionNodeMap & TStatementNodeMap

export type TNode<T extends ENodeType = ENodeType> = TNodeMap[T]

export type TBaseNodeAttr = {
  /** 额外的信息 */
  extra?: {
    /** 是否被小括号包裹 */
    parenthesized?: boolean
    /** 小括号的开始定位信息 */
    parentStart?: TPositionInfo
  }
  /** 定位信息 */
  loc?: {
    start: TPositionInfo
    end: TPositionInfo
  }
}

// 特殊节点定义
export interface IDictionaryProperty extends TBaseNodeAttr {
  type: ENodeType.DictionaryProperty
  key: TExpressionNode
  value: TExpressionNode
}

export interface IAssignmentParam extends TBaseNodeAttr {
  type: ENodeType.AssignmentParam
  name: IIdentifier
  value: TExpressionNode
}

export interface IExceptHandler extends TBaseNodeAttr {
  type: ENodeType.ExceptHandler
  errName?: IIdentifier
  name?: IIdentifier
  body: IBlockStatement
}

// 表达式节点定义
export interface INoneLiteral extends TBaseNodeAttr {
  type: ENodeType.NoneLiteral
}

export interface IBooleanLiteral extends TBaseNodeAttr {
  type: ENodeType.BooleanLiteral
  value: boolean
}

export interface INumberLiteral extends TBaseNodeAttr {
  type: ENodeType.NumberLiteral
  value: number
  raw: string
}

export interface IStringLiteral extends TBaseNodeAttr {
  type: ENodeType.StringLiteral
  value: string
  raw: string
}

export interface IIdentifier extends TBaseNodeAttr {
  type: ENodeType.Identifier
  name: string
}

export interface IUnaryExpression extends TBaseNodeAttr {
  type: ENodeType.UnaryExpression
  oprator: '-' | '+'
  argument: TExpressionNode
}
export interface IIfExpression extends TBaseNodeAttr {
  type: ENodeType.IfExpression
  test: any // todo
  body: any // todo
  alternate: any | IIfExpression // todo
}
export interface ITupleExpression extends TBaseNodeAttr {
  type: ENodeType.TupleExpression
  elements: TExpressionNode[]
}

export interface IArrayExpression extends TBaseNodeAttr {
  type: ENodeType.ArrayExpression
  elements: TExpressionNode[]
}
export interface IDictionaryExpression extends TBaseNodeAttr {
  type: ENodeType.DictionaryExpression
  properties: IDictionaryProperty[]
}
export interface IBinaryExpression extends TBaseNodeAttr {
  type: ENodeType.BinaryExpression
  operator: string
  left: any // todo
  right: any // todo
}

export interface IVariableDeclaration extends TBaseNodeAttr {
  type: ENodeType.VariableDeclaration
  kind: string
  declarations: IIdentifier[]
}

export interface IAssignmentExpression extends TBaseNodeAttr {
  type: ENodeType.AssignmentExpression
  targets: (IIdentifier | IMemberExpression | ITupleExpression)[]
  value: any // todo
}

export interface ISliceExpression extends TBaseNodeAttr {
  type: ENodeType.SliceExpression
  lower: TExpressionNode
  upper: TExpressionNode
  step: TExpressionNode
}

export interface IMemberExpression extends TBaseNodeAttr {
  type: ENodeType.MemberExpression
  object: IIdentifier | IStringLiteral | IMemberExpression | ICallExpression
  property: IIdentifier | INumberLiteral | IStringLiteral | ISliceExpression
}

export interface ICallExpression extends TBaseNodeAttr {
  type: ENodeType.CallExpression
  callee: IIdentifier | IMemberExpression | ICallExpression
  params: TExpressionNode[]
  keywords: IAssignmentParam[]
}

export interface IImportStatement extends TBaseNodeAttr {
  type: ENodeType.ImportStatement
  names: IIdentifier[]
  module?: IIdentifier | IMemberExpression
}

// 语句节点定义
export interface IFunctionDeclaration extends TBaseNodeAttr {
  type: ENodeType.FunctionDeclaration
  id: IIdentifier
  params: IIdentifier[]
  defaults: TExpressionNode[]
  body: IBlockStatement
}

export interface IClassDeclaration extends TBaseNodeAttr {
  type: ENodeType.ClassDeclaration
  id: IIdentifier
  bases: IIdentifier[]
  keywords: IAssignmentParam[]
  body: IBlockStatement
}

export interface IBlockStatement extends TBaseNodeAttr {
  type: ENodeType.BlockStatement
  body: (TExpressionNode | TStatementNode)[]
}

export interface IEmptyStatement extends TBaseNodeAttr {
  type: ENodeType.EmptyStatement
}

export interface IIfStatement extends TBaseNodeAttr {
  type: ENodeType.IfStatement
  test: any // todo
  body: IBlockStatement
  alternate?: IBlockStatement | IIfStatement
}

export interface ITryStatement extends TBaseNodeAttr {
  type: ENodeType.TryStatement
  body: IBlockStatement
  handlers?: IExceptHandler[]
  finalBody: IBlockStatement
}

export interface IProgram extends TBaseNodeAttr {
  type: ENodeType.Program
  body: (TExpressionNode | TStatementNode)[]
}